refac: Improve python module structure
