+++ source/databricks/calculation-engine/tests/integration/basis_data/test_get_master_basis_data_df.py
@@ -14,7 +14,7 @@
 
 import pytest
 from package.codelists import MeteringPointType, MeteringPointResolution
-from package.basis_data import get_master_basis_data_df
+from package.calculation_input.basis_data import get_master_basis_data_df
 from datetime import datetime
