+++ source/databricks/calculation-engine/tests/integration/calculation_output/output_writers/test_basis_data_writer.py
@@ -26,7 +26,7 @@ from package.codelists import (
     TimeSeriesQuality,
 )
 from package.constants import Colname
-from package.calculation_output.output_writers import BasisDataWriter
+from package.calculation_output.basis_data_writer import BasisDataWriter
 from pyspark.sql import DataFrame, SparkSession
 from tests.helpers.assert_calculation_file_path import (
     CalculationFileType,
